# Conflicts:
#	Assets/Amanita/Editor/Scripts/Variables/VariableDataDrawer.cs
#	Assets/Amanita/Resources/GuidRegistries.meta
#	Assets/Amanita/Scripts/Components/AmanitaManager.cs
#	Assets/Amanita/Systems/NewSaveSys/Commands/SaveToSlot.cs
#	Assets/Amanita/Systems/NewSaveSys/Core/SaveSystem.cs
#	Assets/Amanita/Systems/NewSaveSys/EventHandlers/SaveSlotSelectedEvent.cs
#	Assets/Amanita/Systems/NewSaveSys/UI/SaveSlotViewComposer.cs
#	Assets/Amanita/Systems/NewSaveSys/UI/Views.meta
#	Assets/FungusExamples/NewSaveSystem/Scenes/newSaveSystemDemoScene.unity
#	Assets/Saves/saveData_06.save.bak.meta
#	Assets/TempSaves.meta
#	Assets/TestSaves.meta
#	Assets/Tests/PlayMode/SaveSys/CommonTestFunctionality.cs

+++ Assets/Amanita/Editor/Scripts/FlowchartWindow/BlockClipboard.cs
@@ -52,7 +52,7 @@ namespace Amanita.VScripting.EditorUtils
             return result;
         }
 
-        public virtual bool HasMultiEntriesWithIDs(IList<int> ids)
+        public virtual bool HasMultiEntriesWithIDs(IList<ushort> ids)
         {
             bool result = true;
